+++ crates/chat_ui/src/actions.rs
@@ -1,7 +1,6 @@
 use gpui::Action;
 use nostr_sdk::prelude::*;
 use serde::Deserialize;
-use settings::SignerKind;
 
 #[derive(Action, Clone, PartialEq, Eq, Deserialize)]
 #[action(namespace = chat, no_json)]
@@ -9,7 +8,6 @@ pub enum Command {
     Find,
     Insert(&'static str),
     ChangeSubject(String),
-    ChangeSigner(SignerKind),
     ToggleBackup,
     Copy(PublicKey),
     ViewProfile(PublicKey),