+++ desktop/src/main.rs
@@ -76,11 +76,6 @@ fn main() {
                 // Initialize person registry
                 person::init(window, cx);
 
-                // Initialize device signer
-                //
-                // NIP-4e: https://github.com/nostr-protocol/nips/blob/per-device-keys/4e.md
-                device::init(window, cx);
-
                 // Initialize app registry
                 chat::init(window, cx);